Refactor into a proper module
* rename everything to [PEP-8 compliant module names](https://pep8.org/#package-and-module-names)
* provide setup.py for local and global installation
* provide setup instructions
* move documentation for commands to docs/ folder
* improve README
* move visualization example to subfolder, it is not really a "script",
  but an example usage
