+++ obs/bin/obs_filter_privacy.py
