+++ obs/face/annotate/BeliefPropagationChain.py
