+++ obs/face/mapping/__init__.py
@@ -1,0 +1,5 @@
+from .Way import Way
+from .LocalMap import LocalMapTangential, Gnomonic, AzimuthalEquidistant
+from .RoadContainerAABBtree import RoadContainerAABBtree
+from .Roads import Roads
+
