+++ obs/face/osm/__init__.py
@@ -1,0 +1,1 @@
+from .DataSource import DataSource
